userspace: provide alternative IO API
Using NBD as an IO transport channel is less than ideal, considering
that the WinSock Kernel (WSK) doesn't perform very well, especially
on Windows Server 2016.

We're adding an alternative transport. DeviceIoControl calls will be used
for fetching IO requests and sending asynchronous IO replies.

For now, we're going to call it WVBD (Windows Virtual Block Device).

This commit adds an userspace library that will make it easier to
consume this API. For convenience, we'll provide a high level API along
taking care of the dispatching boilerplate, as well as a low level API (the
*Ioctl* functions).

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>
Signed-off-by: Alin Gabriel Serdean <aserdean@cloudbasesolutions.com>