driver: update WNBD API
We're updating the WNBD driver interface, adding an alternative
to the NBD protocol, which doesn't perform well enough (mostly
due to WSK, especially on Windows Server 2016).

The new interface is also meant to be more consistent, easier to
extend and less NBD centric.

This commit updates the driver, implementing the new interface,
also using the new structures internally.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>